// frontend/src/app/generated/model/userRegistrationRequest.ts
/**
 * MDMA_API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserRegistrationRequest { 
    username: string;
    email: string;
    password: string;
    firstName?: string;
    lastName?: string;
}